+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ## Unreleased: mitmproxy next
 
+- Fix a race condition when updating the flow list in mitmweb.
 
 ## 24 May 2025: mitmproxy 12.1.0